Extract PackageTypingsGenerator._followAliases() into a separate class SymbolAnalyzer
